requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorlog==6.7.0
contourpy==1.1.1
cycler==0.11.0
datasets==2.14.5
dill==0.3.7
dnspython==2.4.2
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
hikari==2.0.0.dev120
hikari-lightbulb==2.3.3
huggingface-hub==0.17.3
idna==3.4
kiwisolver==1.4.5
loguru==0.7.2
matplotlib==3.8.0
multidict==6.0.4
multiprocess==0.70.15
numpy==1.26.0
openai==0.27.8
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pinecone-client==2.2.4
plotly==5.17.0
pyarrow==13.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
scipy==1.11.2
six==1.16.0
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.65.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
xxhash==3.3.0
yarl==1.9.2
Flask==2.1.0
gunicorn==20.1.0
google-cloud-secret-manager==2.10.0
hikari-miru==3.2.2

